function done

+++ garo/inc/functions.h
@@ -3,8 +3,6 @@
 * 
 * author: Garo Demergian
 * created: 2020-01-06
-* notes: 
-* 
 * 
 * ver: 2020-01-06 first version
 * 
@@ -16,17 +14,13 @@
 using namespace std;
 
 void welcomeMSG();
-int getAndValidateInput();
+double getAndValidateInput();
 int getAndValidateChoice(int validNumber);
-int getCircumference();
-int getArea();
-int getVolume();
-int getFirstEdge();
-int getSecondEdge();
-int getHeight();
-
-
-
-
+double getCircumference();
+double getArea();
+double getVolume();
+double getFirstEdge();
+double getSecondEdge();
+double getHeight();
 
 #endif
